Refactor chart plugins and label logic
Extracted chart plugin logic and label index calculation into reusable utility modules (plugins.js and labels.js). Updated AreaLineChart, BarLineChart, LineChart, and ScatterChart components to use these utilities, improving code maintainability and reducing duplication.
